refactor: rename file to reflect data returned
